Went through the Wall and pedantic on MacOs 10.9

+++ reg-apps/reg_average.cpp
@@ -74,7 +74,7 @@ int main(int argc, char **argv)
       return 0;
    }
    // Check if help is required
-   for(size_t i=1; i<argc; ++i)
+   for(int i=1; i<argc; ++i)
    {
       if(strcmp(argv[i],"-h")==0 ||
             strcmp(argv[i],"-H")==0 ||
@@ -89,7 +89,7 @@ int main(int argc, char **argv)
    }
    // Command line
    printf("\nCommand line:\n\t");
-   for(size_t i=0; i<argc; ++i)
+   for(int i=0; i<argc; ++i)
       printf("%s ",argv[i]);
    printf("\n\n");
 
@@ -314,7 +314,7 @@ int main(int argc, char **argv)
          size_t affineNumber = (argc - 4)/2;
          // All affine matrices are read in
          mat44 *affineMatrices = (mat44 *)malloc(affineNumber*sizeof(mat44));
-         for(size_t i=4, j=0; i<argc; i+=2,++j)
+         for(int i=4, j=0; i<argc; i+=2,++j)
          {
             if(reg_isAnImageFileName(argv[i]))
             {
@@ -379,7 +379,7 @@ int main(int argc, char **argv)
          tempImage->scl_inter=0.f;
          tempImage->data = (void *)malloc(tempImage->nvox*tempImage->nbyper);
          // warp all floating images and sum them up
-         for(size_t i=5, j=0; i<argc; i+=2,++j)
+         for(int i=5, j=0; i<argc; i+=2,++j)
          {
             nifti_image *floatingImage = reg_io_ReadImageFile(argv[i]);
             if(floatingImage==NULL)
@@ -444,7 +444,7 @@ int main(int argc, char **argv)
          sprintf(msg,"reg_average: Number of input transformations: %i",(argc-4)/operation);
          reg_print_msg_debug(msg);
 #endif
-         for(size_t i=(operation==2?4:5); i<argc; i+=operation)
+         for(int i=(operation==2?4:5); i<argc; i+=operation)
          {
             nifti_image *transformation = reg_io_ReadImageFile(argv[i]);
             if(transformation==NULL)
@@ -557,7 +557,7 @@ int main(int argc, char **argv)
          nifti_image *tempImage = nifti_copy_nim_info(averageImage);
          tempImage->data = (void *)malloc(tempImage->nvox*tempImage->nbyper);
          // Iterate over all the transformation parametrisations
-         for(size_t i=(operation==2?4:5); i<argc; i+=operation)
+         for(int i=(operation==2?4:5); i<argc; i+=operation)
          {
             nifti_image *transformation = reg_io_ReadImageFile(argv[i]);
             if(transformation==NULL)
